requirements.txt
torch
torchvision
pandas
numpy
Pillow
opencv-python
streamlit
scikit-learn
tqdm
reportlab
